requirements.txt
# 基础依赖
python-dotenv>=1.1.0
requests>=2.32.3

# 数据处理
pandas==2.1.4
openpyxl==3.1.2

# LangChain相关
langchain>=0.3.26
langchain-core>=0.3.66
langchain-community>=0.3.27
langgraph>=0.2.0
langsmith>=0.1.0
openai>=1.86.0

# 向量存储
chromadb==0.4.15
protobuf<6,>=3.20
rich<14,>=10.14.0

# 文档处理
python-docx==1.0.1
PyPDF2==3.0.1

# 结构化PDF处理增强
pdfplumber>=0.9.0
pypdf>=3.0.0
pdf2image>=1.16.0

# Web界面
Flask>=2.3.0